// app/src/api/forum/ForumClient.ts
export interface Post {
  id: string;
  author: string;
  title: string;
  body: string;
  post_type: string;   // "thesis" | "post-mortem" | "general"
  created_at: number;
}